+++ src/muncher/items.js
@@ -4,6 +4,7 @@ import { munchNote, getCampaignId, download } from "./utils.js";
 import getInventory from "../parser/inventory/index.js";
 import utils from "../utils.js";
 import { getCobalt } from "../lib/Secrets.js";
+import logger from "../logger.js";
 
 async function getCharacterInventory(items) {
   return items.map((item) => {
@@ -139,7 +140,9 @@ export async function parseItems() {
   const uploadDirectory = game.settings.get("ddb-importer", "other-image-upload-directory").replace(/^\/|\/$/g, "");
 
   // to speed up file checking we pregenerate existing files now.
+  logger.info("Checking for existing files...");
   await utils.generateCurrentFiles(uploadDirectory);
+  logger.info("Check complete, getting ItemData.");
   const results = await getItemData();
   let items = results.items;